% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.monotone.R
\name{is.monotone}
\alias{is.monotone}
\title{Check monotonicity}
\usage{
is.monotone(X)
}
\arguments{
\item{X}{a numerical vector to be tested.}
}
\value{
Returns True or False.
}
\description{
This internel function determines if a numeric vector X is monotone (increasing or decreasing) or not. Ties in
the sequence are considered as monotone.
}
\examples{

a <- 1:10 ;
b <- c( 1:5, 4:2 ) ;
c <- c( 1:4, 4, 4, 4, 5:10 ) ;
d <- c( 10:6, 3:-2 ) ;

is.monotone( a )
is.monotone( b )
is.monotone( c )
is.monotone( d )

}

