% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayo.R
\docType{data}
\name{mayo}
\alias{mayo}
\title{Example data: Mayo Data}
\format{
A data frame with 312 observations and 4 variables:

\code{time}: event time or censoring time

\code{censor} : censoring indicator.

\code{mayoscore4} and \code{mayoscore5}: derived from 4 and 5 covariates respectively.
}
\usage{
data(mayo)
}
\description{
This example dataset is included for illustration. The Mayo PBC data is publicly
   available and has been used in many statistical researches (e.g., Zheng and Heagerty 2005).
   This example dataset is a subset of the full PBC data.
}
\examples{
data(mayo)
head(mayo)
}
\references{
Heagerty, P. J., & Zheng, Y. (2005). Survival model predictive accuracy and ROC curves. Biometrics, 61(1), 92-105.
}
\keyword{datasets}
