% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdBrier.R
\name{tdBrier}
\alias{tdBrier}
\title{Calculate the Brier Score}
\usage{
tdBrier(W, X)
}
\arguments{
\item{W}{a numerical vector of weight estimated by nonparametric weight adjustments (Li \emph{et al.}, 2015). Same length with \code{X}.}

\item{X}{a numerical vector of risk score values. Same length with \code{W}.}
}
\value{
Brier Score as a numerical scalar.
}
\description{
This function reads in a vector of estimated weight and same length biomarker to calculate
   Brier Score by formula (Wu and Li, 2018). This function is used internally by other functions
   in this package.
}
\note{
This function estimates brier score by using the formula from Wu and Li, 2018.
}
\keyword{internal}
